LayoutTests/imported/w3c/web-platform-tests/infrastructure/server/order-of-metas.any-expected.txt

CONSOLE MESSAGE: line 1: ReferenceError: Can't find variable: test
CONSOLE MESSAGE: line 1: ReferenceError: Can't find variable: test
CONSOLE MESSAGE: line 2: ReferenceError: Can't find variable: test

FAIL order of scripts assert_array_equals: value is undefined, expected array
